🎨 move coilset field things to field
